+++ src/Smartstore.Web/Areas/Admin/Models/Settings/CatalogSettingsModel.cs
@@ -289,6 +289,9 @@ namespace Smartstore.Admin.Models
         [LocalizedDisplay("*ShowProductTags")]
         public bool ShowProductTags { get; set; }
 
+        [LocalizedDisplay("*ShowRewardPointsInProductDetail")]
+        public bool ShowRewardPointsInProductDetail { get; set; }
+
         #endregion
 
         public PriceSettingsModel PriceSettings { get; set; } = new();