+++ src/Smartstore.Web/Models/Catalog/ProductReviewsModel.cs
@@ -9,6 +9,8 @@ namespace Smartstore.Web.Models.Catalog
         public int RatingSum { get; set; }
         public int TotalReviews { get; set; }
         public bool AllowCustomerReviews { get; set; }
+        public int RewardPoints { get; set; }
+        public Money? RewardPointsAmount { get; set; }
     }
 
     [LocalizedDisplay("Reviews.Fields.")]
@@ -18,7 +20,7 @@ namespace Smartstore.Web.Models.Catalog
         public LocalizedValue<string> ProductName { get; set; }
         public string ProductSeName { get; set; }
         public int TotalReviewsCount { get; set; }
-        public List<ProductReviewModel> Items { get; set; } = new();
+        public List<ProductReviewModel> Items { get; set; } = [];
         public bool ShowVerfiedPurchaseBadge { get; set; }
         public bool IsReviewsDetailPage { get; set; }