+++ changelog.md
@@ -46,8 +46,10 @@
 - Reverse proxy: added support for `X-Forwarded-Prefix` header (for `PathBase` resolution)
 - Page Builder
 	- Added **AudioPlayer** block
+- Product detail page
+  - #997 Added setting to disable display of product tags on the product detail page.
+  - #1127 Display reward points in product detail.
 - #858 Implemented Paypal package tracking
-- #997 Added setting to disable display of product tags on the product detail page  
 - #1100 Display customer generic attributes in backend
 - #1129: Extend the PrivacySettings CookieConsentRequirement option to include Switzerland when choosing the option RequiredInEUCountriesOnly
 - #762 CookieManager: Scripts are now loaded immediately after consent without refreshing the page.